--- benchmarks/go.mod
@@ -1,33 +1,0 @@
-module github.com/tarmac-project/hord/benchmarks
-
-go 1.21
-
-toolchain go1.22.4
-
-require (
-	github.com/tarmac-project/hord v0.4.0
-	github.com/tarmac-project/hord/drivers/bbolt v0.4.0
-	github.com/tarmac-project/hord/drivers/cassandra v0.4.0
-	github.com/tarmac-project/hord/drivers/hashmap v0.4.0
-	github.com/tarmac-project/hord/drivers/nats v0.4.0
-	github.com/tarmac-project/hord/drivers/redis v0.4.0
-)
-
-require (
-	github.com/FZambia/sentinel v1.1.1 // indirect
-	github.com/gocql/gocql v1.6.0 // indirect
-	github.com/golang/snappy v0.0.4 // indirect
-	github.com/gomodule/redigo v1.9.2 // indirect
-	github.com/hailocab/go-hostpool v0.0.0-20160125115350-e80d13ce29ed // indirect
-	github.com/klauspost/compress v1.17.7 // indirect
-	github.com/madflojo/hord v0.3.0 // indirect
-	github.com/nats-io/nats.go v1.36.0 // indirect
-	github.com/nats-io/nkeys v0.4.7 // indirect
-	github.com/nats-io/nuid v1.0.1 // indirect
-	github.com/rogpeppe/go-internal v1.12.0 // indirect
-	go.etcd.io/bbolt v1.3.10 // indirect
-	golang.org/x/crypto v0.20.0 // indirect
-	golang.org/x/sys v0.17.0 // indirect
-	gopkg.in/inf.v0 v0.9.1 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
